:sparkles: GroupDataSource 추상화
