+++ data/src/main/java/com/whyranoid/data/di/GroupModule.kt
@@ -1,5 +1,7 @@
 package com.whyranoid.data.di
 
+import com.whyranoid.data.group.GroupDataSource
+import com.whyranoid.data.group.GroupDataSourceImpl
 import com.whyranoid.data.group.GroupRepositoryImpl
 import com.whyranoid.domain.repository.GroupRepository
 import dagger.Binds
@@ -10,6 +12,9 @@ import dagger.hilt.components.SingletonComponent
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class GroupModule {
+
+    @Binds
+    abstract fun bindGroupDataSource(groupDataSourceImpl: GroupDataSourceImpl): GroupDataSource
 
     @Binds
     abstract fun bindGroupRepository(groupRepositoryImpl: GroupRepositoryImpl): GroupRepository
